+++ app/testblog/forms.py
@@ -1,6 +1,6 @@
 from trionyx import forms
 
-from app.testblog.models import Tag
+from app.testblog.models import Tag, Post
 
 
 @forms.register(code='new', default_edit=True)
@@ -10,3 +10,10 @@ class TagForm(forms.ModelForm):
     class Meta:
         model = Tag
         fields=['name']
+
+
+@forms.register(default_edit=True)
+class PostForm(forms.ModelForm):
+    class Meta:
+        model = Post
+        fields = ['title', 'content', 'publish_date', 'category', 'status', 'price', 'tags']