+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Added
 ~~~~~
 - Add create_reusable_app command
 - Add ProgressBar component
+- Add Unordered and Ordered list components
 
 
 [2.0.1] - 19-12-2019